Add file via upload
The file allows users to add or remove projects and milestones.
This error is caused by the current limitations of the Solidity compiler, which does not support directly copying a memory array (Milestone memory[]) to storage. This error usually occurs when attempting to copy a Project struct from storage to memory. To fix this, create a new project as Project memory newProject and push it to storage using projects[msg.sender].push(newProject);.